fix(mixin): pod selector regex for deployments (#4603)
The selector used in the mixin for pods started with (.*-mimir-). However, using this helm chart in a wrapper chart name mimir and deploying the helm chart with release name mimir this regex would fail to find any pods. This PR changes the regex to (.*mimir-) so that helm releases with the name mimir will also function.

---------

Signed-off-by: David van der Spek <vanderspek.david@gmail.com>