store-gateway: read series from bucket with io.Reader (#4926)
Uses an io.Reader to read series data from the bucket, so we never hold everything in a buffer. It also introduces pooling for same bytes.

This PR introduces a reader offsetTrackingReader which can keep track of its offset in a buffer. I plan to introduce this reader in loadChunks as well to simplify the code there a bit.


---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>