Refactor analyzer (#5180)
* Mimirtool: refactor grafana and ruler analyzer

This commit moves the computation of the `MetricsUsed` in
AnalyzeGrafana/AnalyzeRuler. Doing so facilitate usage via vendoring in
downstream applications.

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>

* Mimirtool: refactor prometheus analyzer

This commit refactor prometheus analyzer to facilitate use via vendoring
in downstream applications. The may reason why the refactoring is
useful is that it split the metric analysis from the interaction on the
filesystem. As such, `IdentifyMetricsInPrometheus` can be used without
touching the filesystem.

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>

---------

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>