Use three ingesters for mimir-microservices-mode docker compose (#5941)
Update the mimir-microservices-mode docker compose environment to use three
ingesters instead of two. We're already running two ingesters so it's not a
huge burden to run another one. This allows us to use the default replication
factor and easily test that the write path is robust against a single ingester
crashing.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>