query-frontend: remove duplicate defaulting of instant query `time` param (#7026)
After 6985 the default value of the time parameter (i.e. time.Now()) is set in the PrometheusCodec. This PR removes the
default set in an HTTP middleware.

With this change the time parameter in the HTTP request is no longer populated if
the client didn't provide it. Instead, it will only be populated in the decoded `querymiddleware.Request`. The HTTP parameter was referenced in a few places
* [DecodeInstantQueryTimeParams](https://github.com/grafana/mimir/blob/ad1953d5da1f85dc2814e65d8abcf3eb4d841614/pkg/frontend/querymiddleware/codec.go#L295) which now also falls back to the default (after PR 6985). This is also used for partial queries, but those shouldn't observe an unset parameter (since the parent query had the parameter defaulted to `time.Now()`).
* invocations of [ParseTimeParam](https://github.com/grafana/mimir/blob/f83ace2ce367174fd6789f9ab8e74a8d793de6b0/pkg/util/time.go#L40). One of them is in DecodeInstantQueryTimeParams, the other two are for label values requests, which have their own logic for defaulting the parameters.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>