Set -server.grpc.num-workers=100 by default (#7131)
* Set -server.grpc.num-workers=100 by default

This sets the `-server.grpc.num-workers` to 100 by default, and changes
the category of that flag to `advanced`.

This is the value we've tested in Grafana Cloud for a while. We've
observed an improvement of around 5% in ingester CPU compared to running
without this flag set.

Note that this is not a limit, but the number of pre-allocated
long-lived workers. When all workers are busy, a new goroutine will be
spawned (previous default behaviour).

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>

* Update CHANGELOG.md

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>

---------

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>