Alertmanager: Increase likehood of useful results from /grafana/receivers. (#8540)
* Alertmanager: Increase likehood of useful results from /grafana/receivers.

The experimental, extended /receivers API provides information about the last
notification attempt for receivers, including the last error and duration.

Currently, because of how the Alertmanager ring is configured, we only require
a successful response from one replica. For the receivers API however this is
insufficient, because any of the 3 replicas could have sent notifications
recently. It's more likely that one of the replicas will (the replica in
"position 0"), but any replica could send any given notification, so asking
all replicas is arguably more robust.

* Review comments