store-gateway: dont leak goroutines when preloading (#8546)
I discovered this while working on XXX. This shouldn't be a bug today because nothing above the preloading iterator looks at the context. But it can become a problem in the future. I also verified in a few production clusters at GL and couldn't spot a gradual increase in the number of goroutines.

It's blocking detecting leaking goroutines in my work for XXX, so I'd like to fix it first.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>