Allow namespaces to be protected and deny modification unless specified (#8444)
* Allow namespaces to be protected and deny modification unless specified

This is a draft for allowing Mimir to have a concept of protected namespaces. There are two key design pieces:

1. The use of `X-Mimir-Ruler-Override-Namespace-Protection` as a header to override the protection and allow modification e.g. `X-Mimir-Ruler-Override-Namespace-Protection: mynamespace,myothernamespace`
2. The use of `X-Mimir-Ruler-Protected-Namespaces` as a header from Mimir itself to tell the client that this namespace is actually protected whilst it's being read e.g. `X-Mimir-Ruler-Protected-Namespaces: mynamespace,myothernamespace`

It's important to note that both header values can be comma-separated list of strings. In the case of `X-Mimir-Ruler-Protected-Namespaces` is to list _all_ the protected namespaces on a list request and on the case of `X-Mimir-Ruler-Override-Namespace-Protection` is to include all namespaces this client is allowed to modify.

Signed-off-by: gotjosh <josue.abreu@gmail.com>
---------
Signed-off-by: gotjosh <josue.abreu@gmail.com>
Co-authored-by: taylor c <41653732+tacole02@users.noreply.github.com>