jsonnet: allow rollout-operator to be used as webhook endpoint (#8743)
* jsonnet: allow rollout-operator to be used as webhook endpoint

This change adds jsonnet configuration that allows the rollout-operator to be
used as a validating or mutating webhook for changes to statefulsets. This is
required to use the `no-downscale` or `prepare-downscale` labels on stateful
components.

This doesn't make any changes to the functionality used internally at Grafana
or make anything more configurable. For example, this jsonnet still has the
rollout-operator create self-signed certificates for the HTTPS webhook
endpoints.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>

* Ignore webhook labels for services

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>

---------

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>