distributor: promote Retry-After header to stable, simplify config (#8694)
Preparing for the 2.14 release. We've been running this in production since it was introduced and there haven't been any problems. The results have been positive in reducing the thundering herd retry behaviour.

Before promoting to stable we also wanted to simplify the configuration. After the discussion in this PR we stopped at replacing `-distributor.retry-after-header.max-backoff-exponent` and `-distributor.retry-after-header.base-seconds` with `.min-backoff` and `.max-backoff` for easier configuration.

Now the Retry-After value is calculated as `min_backoff * 2^(retry_attempt - 1)` with a random jitter of 50% in either direction and capped to `max_backoff`