ingest: use single configuration for fetch concurrency (#10156)
This replaces the two previous flags with one single one.

```
# old flags
-ingest-storage.kafka.startup-fetch-concurrency
-ingest-storage.kafka.ongoing-fetch-concurrency

# new flag
-ingest-storage.kafka.fetch-concurrency-max
```

At GL we've been running with a single value for both. The reason for having two flags in the first place is because we couldn't find a balance between throughput and latency. This was largely fixed with the change to not issue fetch requests for beyond the HWM (9892). So now we no longer need the complexity of being able to change concurrency settings

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>