Deflake test `TestDistributor_LabelValuesCardinality` (#11095)
This fixes the polling bit in this test that was intended to retry when not all mockIngesters had yet received the push. The problem was the initial asserts inside the poll had failed while the ultimate return value of the polled function did match the expected values. So the Poll would exit with a failed assertion instead of retrying.

Would fail 1-5x per hundred runs before the fix.