Track cortex_ingest_storage_writer_latency_seconds for failures too (#11770)
#### What this PR does

During a recent incident, I've learned that
`cortex_ingest_storage_writer_latency_seconds` is very misleading if
records production times out, because this metric only tracks the
latency for successfully produced records. In this PR I'm changing it to
track latency both on success and error, but distinguishing between the
two cases using a `outcome` label.

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>