writes dashboard: add Influx write requests row (#11731)
#### What this PR does

The new Influx endpoint in the distributor is effectively a sibling of
the OTLP endpoint, and so I have made similar additions to those that
exist for the OTLP endpoint.

This PR adds a new row to 'Writes' Dashboard showing the stats for
metrics arriving by the Influx endpoint. (Based heavily on the existing
instrumentation for the OTLP endpoint.) This adds the appropriate routes
to the queries and to the SLO based queries.

Also adds and entry for `err-mimir-distributor-max-influx-request-size`
to the runbook.

Documentation for the Influx endpoint in the API reference (akin to
https://grafana.com/docs/mimir/latest/references/http-api/#otlp) to be
added in a separate PR.

It's designed look exactly like the `Gateway - OTLP write requests` row
in the current dashboard. The new row looks like this (data is from
`cortex-dev-01` as that is the only cell that has traffic (Snythethic
Monitoring) to this cell.


![image](https://github.com/user-attachments/assets/ee289245-eb4b-4ed1-b7ef-ef714da232b6)

#### Checklist

- [ ] ~Tests updated.~
- [x] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
~[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.~

---------

Signed-off-by: alexgreenbank <alex.greenbank@grafana.com>